[#157903353] convert hungry animation to react
