+++ .eslintrc
@@ -7,11 +7,18 @@
             "jsx": true
         }
     },
+    "parser": "babel-eslint",
+    "plugins": [ "babel" ],
     "rules": {
+        "babel/new-cap": 1,
+        "babel/no-invalid-this": 1,
+        "babel/semi": 1,
+        "babel/no-unused-expressions": 1,
         "import/no-unresolved": "off",
         "import/extensions": "off",
         "consistent-return": "off",
         "linebreak-style": "off",
+        "no-return-assign": "off",
         "comma-dangle": ["error", {
             "arrays": "never",
             "objects": "never",
