+++ client/src/components/Home/home.scss
@@ -1,0 +1,11 @@
+.hungry{
+  font-style: normal !important;
+  margin: 3px;
+  cursor: pointer;
+  transition: font-size 1s
+  
+}
+
+.hungry.bounceIn{
+    font-size: 40px;
+}
