+++ package.json
@@ -56,7 +56,6 @@
     "eslint-config-airbnb": "^16.1.0",
     "eslint-plugin-import": "^2.11.0",
     "eslint-plugin-jsx-a11y": "^6.0.3",
-    "eslint-plugin-react": "^7.7.0",
     "express": "^4.16.3",
     "express-group-routes": "^1.1.0",
     "express-validation": "^1.0.2",
@@ -106,6 +105,7 @@
     "winston-transport": "^4.0.0"
   },
   "devDependencies": {
+    "babel-eslint": "^8.2.3",
     "babel-plugin-add-module-exports": "^0.2.1",
     "babel-plugin-istanbul": "^4.1.6",
     "babel-plugin-transform-class-properties": "^6.24.1",
@@ -114,6 +114,8 @@
     "babel-plugin-universal-import": "^2.0.0",
     "babel-preset-env": "^1.6.1",
     "babel-preset-react": "^6.24.1",
+    "eslint-plugin-babel": "^5.1.0",
+    "eslint-plugin-react": "^7.9.1",
     "webpack-cli": "^2.1.5"
   }
 }
